Completed "Dodge-em" exercise! See note below.
I am overall happy with the result, and surprised with the power of p5!  If I were to revise the program (which I hope to do), I would add a conditional statement that would make the monsters bounce off of one another. This would add further game difficulty for the user.