Progressing on the Acorn Catcher project. Please see note below
The basic fundamentals of the game have been created and are functioning correctly. Next, I must make the sate change to "Winner" once the user obtains the desired amount of points. I also need to constrain mouseX to the canvas, prevent it from moving vertically on the Y-axis, and add a delayed acceleration for added difficulty!