Progressing on the Acorn Catcher project. Please see note below
My vision is almost complete! I have added background music in the "Simulation" and "Winner" state that fits the overall theme of the game. I am also in the process of changing the shapes (ellipse) to character images, and applying background images to the "Simulation" state. All that is left is  to organize / comment my code.