Progressing on EX4: Age of Aquariums exercise. See note below.
I have decided to remain relatively content with this program. I started with the fish simulation, and have added a shark in which the user will control with mouseX and mouseY. The objective will be to eat 500 points worth of fish within 20 seconds. So far the programming is going alright. I have stumbled upon an error when attempting to target an objects variables inside of an array.... My question has been sent on the discord!