Progressing on EX4: Age of Aquariums exercise. See note below.
The program is technically complete and functions as intended. Since we were allotted until Sunday night to hand in the assignment, I have decided that I will finish organizing functions and commenting over the weekend. Perhaps a thought or idea will will come to mind, that may want to incorporate!