E5: Started "juggle-garden" Exercise. See note below.
I have decided to add complexity to the juggle simulation, with the introduction to a token system where the users objective is direct the balls into the tokens to progress through the simulation. Thus far, I have created a "Token" class and have been successful at displaying the tokens on screen. I still need to add states, apply noise to the tokens and write a function to say that when the ball hits the token, the token is too disappear. When all the tokens disappear the user wins.