Completed EX4: juggle-garden exercise. See note below.
I am overall happy with this program and grateful for Madeline's assistance :) The program may not be as visually pleasing as some other, however I truly believe it is my most advanced  program due to it's technical complexity. It uses OOP, various loops, arrays, classes, functions etc. This program will act as a great reference for future work.