Project 2: Implementation of "ballPuzzle". See notes below
My second puzzle is based off my creation from the exercise "Juggle-Garden". My idea was to first implement the simulation into my project, and then modify as seen necessary. The next step is to add an auditory timer, delay to the "click" sound, and potentially use the random function to change the colour of the balls upon impact...